Add benchmarking utils in transv.isotrop. material
- Added utilities for benchmarking the time integration algorithm of
InelasticDefgradTransvIsotropElastViscoplast. Specifically, a dedicated
object is used to track the number of substeps, the number of
iterations, the computation times and the different error types that
arise during computation. The functionality is only provided for
single-processor computations.
- A csv writer is integrated within the BenchmarkUtils object, which
writes runtime output for the time integration benchmarking.
- The user can select whether to write dedicated runtime output for the
time integration via a dedicated boolean variable.