Add solid mechanics tutorial
This tutorial provides a structure, which deforms plastically.
Also, the test is added to the framework-tests.

The general structure of the tutorials chapter is slightly changed,
some more information about the preprocessor cubit has moved to an
introductory section.

For some more versatility, the material model gets a new parameter,
TOL, which is given in many other models as well.