Replace erroneous nullptr MPI_COMM initialization
In some cases a MPI_COMM variable was initialized with nullptr which
compiles with openmpi because MPI_COMM typedef is a pointer in openmpi.
For intel-mpi, this however leads to an error because the underlying
type is an int. The correct initialization is MPI_COMM_NULL.

Also remove the constructor LevelSetIntersection(int,bool) because
LevelSetIntersection(MPI_COMM,bool) has the same argument types for
intel-mpi.