Add 1st Python surrogate for rough surface contact
This is a first working version with input file and unittests. Paths and
models are hard-coded so far.

The Python model re-implements a linear constitutive law from C++ in
Python. Both give the same results, so this completes the proof of
concept. Further cleanup of the input as well as the handling of the
Python instance will follow.